+++ src/plugin/context.rs
@@ -23,7 +23,7 @@ use rapier::control::CharacterAutostep;
 use rapier::geometry::DefaultBroadPhase;
 
 #[cfg(doc)]
-use crate::prelude::{ImpulseJoint, JointDescription, MultibodyJoint, RevoluteJoint};
+use crate::prelude::{ImpulseJoint, MultibodyJoint, RevoluteJoint, TypedJoint};
 
 /// The Rapier context, containing all the state of the physics engine.
 #[cfg_attr(feature = "serde-serialize", derive(Serialize, Deserialize))]
@@ -888,7 +888,7 @@ impl RapierContext {
     ///
     /// The angle is computed along the revolute joint’s principal axis.
     ///
-    /// Parameter `entity` should have a [`ImpulseJoint`] component with a [`JointDescription::RevoluteJoint`] variant as `data`.
+    /// Parameter `entity` should have a [`ImpulseJoint`] component with a [`TypedJoint::RevoluteJoint`] variant as `data`.
     pub fn impulse_revolute_joint_angle(&self, entity: Entity) -> Option<f32> {
         let joint_handle = self.entity2impulse_joint().get(&entity)?;
         let impulse_joint = self.impulse_joints.get(*joint_handle)?;