+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and new features. Please have a look at the
 
 - Update to rapier `0.21`.
 - Update to nalgebra `0.33`.
-- `ImpulseJoint` and `MultibodyJoint` now have a more detailed enum `JointDescription` instead of a `GenericJoint`.
+- `ImpulseJoint` and `MultibodyJoint` now have a more detailed enum `TypedJoint` instead of a `GenericJoint`.
 You can still access its inner `GenericJoint` with `.as_ref()` or `as_mut()`.
 - `data` fields from all joints (`FixedJoint`...) are now public, and their getters removed.